Added Update_Module Send Update_Module Request but with update flag.

+++ Prototype/Backend/Program/Module/Main/Module.py
@@ -1,4 +1,6 @@
 import os.path
+import subprocess
+
 from flask import Blueprint, render_template, request, session, redirect, url_for
 import time
 import zipfile
@@ -22,7 +24,7 @@ blueprint = Blueprint('main', __name__, url_prefix="/module")
 
 TESTING = True
 
-def scan_file(in_file, modulename, TableScan= False):
+def scan_file(in_file, modulename, TableScan= False, update=True):
     '''
     Function To Scan a python file for correct syntax & imports.
 
@@ -37,6 +39,7 @@ def scan_file(in_file, modulename, TableScan= False):
         10 - TableName not declared with modulename prefix
         11 - Restriced Module Found in Python Files
         12 - Syntax Error Found in Python Files
+        20 - Missing Datatables
     '''
 
     modules = []
@@ -72,6 +75,16 @@ def scan_file(in_file, modulename, TableScan= False):
         for match in matches:
             if f"{modulename}_" not in match:
                 return on_error(10, "Tablename Doesnt doesn't start with modulename")
+        if update:
+            if os.path.exists(f"Program/DB/Models/{modulename}/{os.path.basename(in_file.name)}"):
+                with open(f"Program/DB/Models/{modulename}/{os.path.basename(in_file.name)}") as ref_file:
+                    alt_lines = ref_file.readlines()
+                    for line in alt_lines:
+                        if line in lines:
+                            lines.remove(line)
+                    if lines != []:
+                        return on_error(20, "Table is missing content on Update, Please ensure datatables are not missing")
+
     in_file.seek(0)
     with open(r'Program\templates\whitelisted_modules.txt') as whitelist:
         lines = whitelist.read()
@@ -84,8 +97,11 @@ def scan_file(in_file, modulename, TableScan= False):
         return 0
 
 
-
 def QueryInsertModule(new_module: Module, test=False):
+
+    """ Function to Improt Module into DB
+        if Module exists delete itself if gotten this far into system
+    """
     #If Module Exists Drop it
     existing_modules = Module.query.filter_by(prefix=new_module.prefix).delete()
     from Program import db
@@ -129,9 +145,23 @@ def get_all_plugins():
 def Hello_World():
     return "<h1> Hello World </h1>"
 
+def front_end_installation(temp_dir, module_name, master_dir, update=False):
+    """ Main Function to Facilitate the installation of Front End Components
+        1. Confirm Main.JS exists and pull export function name
+        2. Write the amended Moduledefs.JS file with new imports
+        3. If Logo File exists rename and push to main app
+        4. Push front-end files to front-end & write the module-defs file
 
+    Returns None
 
-def front_end_installation(temp_dir, module_name, master_dir):
+    On Error:
+            Error 16 - When Updating Module, the Main.JS export function has changed, this must stay the same to protect other modules.
+            Error 15 - Main.JS Does Not Exist
+            Error 14 - Export Function Not Found in Main.JS
+            Error 10 - Module Name is not appended to the front of export function
+    """
+    if os.path.exists(f"{temp_dir}\Main.js") == False:
+        return on_error(15, "Front-End Missing Main.js file")
 
     with open(f"{temp_dir}\Main.js") as MainJS:
         content = MainJS.read()
@@ -143,11 +173,13 @@ def front_end_installation(temp_dir, module_name, master_dir):
             return on_error(10, "Module Name not in main Front-End Function Name")
 
     os.chdir("../")
+
     os.chdir("./Front-End-Current/src/")
     front_end_dir = os.getcwd()
     frontEnd_outdir = rf"{front_end_dir}\modules\{module_name}"
     frontEnd_logodir = rf"{front_end_dir}\logos"
 
+    # Write new Module Def Files
     with open(front_end_dir + r".\\moduledefs.js", "r") as file:
         content = file.read()
         imports = ""
@@ -156,32 +188,58 @@ def front_end_installation(temp_dir, module_name, master_dir):
         module_definitions = module_definitions.splitlines()
         i = 0
         for line in module_definitions:
+            if update and module_name in line:
+                pattern = "(?<=import ).*(?= from)"
+                old_functionName = re.findall(pattern, line)[0]
+                if functionName != old_functionName:
+                    return on_error(16, f"Main.js Export default function name changed, please change back to {old_functionName}")
+                else:
+                    break
             if line == '//IMPORT_END':
                 import_pos = i
             if line == '}':
                 last_pos = i
             i = i + 1
-        module_definitions.insert(import_pos, f'import {functionName} from "./modules/{module_name}/main.js";')
-        module_definitions[last_pos] = module_definitions[last_pos] + ","
-        module_definitions.insert(last_pos + 1, f"    {module_name}: {functionName}") # LEAVE 4 SPACES FOR SYNTAX :3
-        new_content = re.sub(pattern, '', content)
-        new_content = new_content
-        module_definitions.append("//REGEX_END")
-        module_definitions.insert(0, "//REGEX_START")
-    if os.path.exists(f"{master_dir}/Program/Temp_Module/ts1/logo.svg"):
+        if not update:
+            module_definitions.insert(import_pos, f'import {functionName} from "./modules/{module_name}/main.js";')
+            module_definitions[last_pos] = module_definitions[last_pos] + ","
+            module_definitions.insert(last_pos + 1, f"    {module_name}: {functionName}") # LEAVE 4 SPACES FOR SYNTAX :3
+            new_content = re.sub(pattern, '', content)
+            new_content = new_content
+            module_definitions.append("//REGEX_END")
+            module_definitions.insert(0, "//REGEX_START")
+    if os.path.exists(f"{master_dir}/Program/Temp_Module/{module_name}/logo.svg"):
+        if os.path.exists(f"{master_dir}/Program/Temp_Module/{module_name}/{module_name}.svg"):
+            os.remove(f"{master_dir}/Program/Temp_Module/{module_name}/{module_name}.svg")
+
         os.rename(f"{master_dir}/Program/Temp_Module/ts1/logo.svg", f"{master_dir}/Program/Temp_Module/ts1/{module_name}.svg")
-        shutil.move(f"{master_dir}/Program/Temp_Module/ts1/{module_name}.svg", frontEnd_logodir)
 
-    shutil.move(rf"{master_dir}\Program\Temp_Module\Front End", frontEnd_outdir)
+        if os.path.exists(f"frontEnd_logodir/{module_name}.svg"):
+            os.remove(front_end_installation)
+        shutil.move(f"{master_dir}/Program/Temp_Module/{module_name}/{module_name}.svg", frontEnd_logodir)
 
-    with open(front_end_dir + r".\\moduledefs.js", "w") as file:
-        file.writelines(line + '\n' for line in (module_definitions + new_content.splitlines()))
+    if update:
+        shutil.rmtree(frontEnd_outdir)
+    shutil.move(rf"{master_dir}\Program\Temp_Module\{module_name}\Front End", frontEnd_outdir)
+
+    if not update:
+        with open(front_end_dir + r".\\moduledefs.js", "w") as file:
+            file.writelines(line + '\n' for line in (module_definitions + new_content.splitlines()))
 
     os.chdir(master_dir)
     return True
 
-def back_end_installation(API_Files, temp_dir, modulename, backend_outdir):
-    if os.path.exists(backend_outdir):
+def back_end_installation(API_Files, temp_dir, modulename, backend_outdir, update=False):
+    """
+    Function for Back-End installation.
+        1. Check if blueprint directory already exists
+        2. Scan Each file is safe to import into system
+
+    Returns True - On Success
+
+    Error Codes None (See Scan_file Function)
+    """
+    if os.path.exists(backend_outdir) and not update:
         shutil.rmtree(temp_dir)
         return on_error(3, "BluePrint Already Exists")
 
@@ -195,14 +253,23 @@ def back_end_installation(API_Files, temp_dir, modulename, backend_outdir):
     return True
 
 
-def table_installation(TableFiles, temp_dir, modulename, Table_outdir):
-    if os.path.exists(Table_outdir):
+def table_installation(TableFiles, temp_dir, modulename, Table_outdir, update=False):
+    """
+    Function for Table installation.
+        1. Check if Table directory already exists
+        2. Scan Each Table file is safe to import into system
+
+    Returns True - On Success
+
+    Error Codes None (See Scan_file Function)
+    """
+    if os.path.exists(Table_outdir) and not update:
         shutil.rmtree(temp_dir)
         return on_error(2, "Table Already Exists")
 
     for file in TableFiles:
         with open(file) as table_file:
-            scan_result = scan_file(table_file, modulename, True)
+            scan_result = scan_file(table_file, modulename, True, update)
             if scan_result:  # If Scan returns an Error
                 table_file.close()
                 shutil.rmtree(temp_dir)
@@ -210,6 +277,11 @@ def table_installation(TableFiles, temp_dir, modulename, Table_outdir):
     return True
 
 def get_module(prefix):
+    """
+    Get module relevant based on the given module prefix
+
+    returns Module obj on success else None
+    """
     Modules = Module.query.filter(Module.prefix == prefix).all()
     return Modules
 
@@ -229,12 +301,14 @@ def upload_module():
         Error Code 12 - Syntax Error Found in Python Files
         On Success - Return new_module Module As Json
     '''
-    if request.method == 'POST':
+    if request.method in ['POST', 'UPDATE']:
+        update = request.method == 'UPDATE'
+
         master_dir = os.getcwd()
         dl_file = request.files['fileToUpload']
         modulename = dl_file.filename.strip(".zip")
 
-        if get_module(modulename) != []:
+        if get_module(modulename) != [] and not update:
             return on_error(1, f"Module {modulename}, Already Exists")
 
         if splitext(dl_file.filename)[1] != ".zip":
@@ -249,27 +323,34 @@ def upload_module():
         TableFiles = dir_tree(rf"{temp_dir}{modulename}\Tables")
         Table_outdir = rf"Program\DB\Models\{modulename}"
         FrontEndDir = rf"{temp_dir}{modulename}\Front End"
+        logo_path = ""
+        if os.path.exists(f"Program/Temp_Module/{modulename}.svg"):
+            logo_path = f'/logos/{modulename}.svg'
 
         if API_Files != []:
-            back_end_success = back_end_installation(API_Files, temp_dir, modulename, api_outdir)
+            back_end_success = back_end_installation(API_Files, temp_dir, modulename, api_outdir, update=update)
             if back_end_success is not True:
                 return back_end_success
 
         if TableFiles != []:
-            table_success = table_installation(TableFiles, temp_dir, modulename, Table_outdir)
+            table_success = table_installation(TableFiles, temp_dir, modulename, Table_outdir, update=update)
             if table_success is not True:
                 return table_success
 
         API_outdir = rf"Program\Module\{modulename}"
 
-        front_end_success = front_end_installation(FrontEndDir, modulename, master_dir)
+        front_end_success = front_end_installation(FrontEndDir, modulename, master_dir, update=update)
 
         # All Modules are Valid, now move to the correct directories, If they exist
         if API_Files != []:
+            if update:
+                shutil.rmtree("{API_outdir.strip(modulename)}/{modulename}")
             shutil.move(f"{temp_dir}{modulename}\Backend", API_outdir.strip(modulename))
             os.rename(f"{API_outdir.strip(modulename)}\Backend", f"{API_outdir.strip(modulename)}/{modulename}")
 
         if TableFiles != []:
+            if update:
+                shutil.rmtree(Table_outdir)
             shutil.move(f"{temp_dir}{modulename}\Tables", Table_outdir)
             tables = convert_to_imports(dir_tree(Table_outdir))
             from Program.DB.Builder import create_db
@@ -283,12 +364,12 @@ def upload_module():
                 shutil.rmtree(f"Program/DB/Models/{modulename}")
             shutil.rmtree(temp_dir)
             return front_end_success
-
         # Upload Tables to Database
 
         DisplayName = request.values['displayName']
         ModulePass = request.values['modulePass']# TODO When User Auth Done, Encrypt module pass
-        new_Module = create_module(str(modulename),DisplayName, ModulePass, True, f'/logos/{modulename}.svg')
+        new_Module = create_module(str(modulename),DisplayName, ModulePass, True, logo_path)
+        new_Module = create_module(str(modulename),DisplayName, ModulePass, True, logo_path)
         QueryInsertModule(new_Module)
         os.chdir(master_dir)
